+++ shub/apps/users/models.py
@@ -103,6 +103,18 @@ class User(AbstractUser):
     class Meta:
         app_label = 'users'
     
+    def has_create_permission(self):
+        '''has create permission determines if the user (globally) can create
+           new collections. By default, superusers and admin can, along with
+           regular users if UER_COLLECTIONS is Ture. Otherwise, not.
+        '''
+        if self.is_superuser is True or self.is_staff is True:
+            return True
+        if settings.USER_COLLECTIONS is True:
+            return True
+        return False
+
+
     def get_label(self):
         return "users"